split pathService into logService, uvService, and pathService
