+++ src/services/monitorService.ts
@@ -4,7 +4,8 @@ import {
 	getMonitorDefaultPort,
 	getRunSerialPort,
 } from '@services/configService';
-import { ensureXousCorePath, getBaoRunner, resolveBaoPy } from '@services/pathService';
+import { ensureXousCorePath, resolveBaoPy } from '@services/pathService';
+import { getBaoRunner } from '@services/uvService';
 import * as vscode from 'vscode';
 
 let monitorTerm: vscode.Terminal | undefined;
